CAF-2108 Missing file for Job Tracking tests (#19)
* CAF-2108 Removed FS container as Job Tracking Worker does not use storage.  For testing purposes Job Tracking Worker Container references File Storage, (FS) as the health returns true always for File Storage, as opposed to CAF Storage which much always be tests for existence.
Updated JobTrackingWorkerIT.java to use TestNG so the test now get run on Build Machine.
Updated JobTrackingWorkerIT.java testProxiedFailureMessage() testcase as due to change in the core Job Tracking Worker, as Failure is no lonver viewed as failure by the Job Tracking Worker.  This test need updated to reflect this updated functionality.

* CAF-2108 Following code review:
Removal of cfg_test_worker-jobtracking_RabbitConfiguration.  Once the test were fixed this was not required as the data is already held in another test-configs configuration file.
Removed the "fs" from the container name.  Going forward the Jobtracking Worker will have only 1 container as it does not use storage.  Therefore no need to provide a CS and a FS variant.

* CAF-2108 Removing redundant link to CAF Builder Jenkins

* CAF-2108 Separating out Rabbit Configuration and Rabbit Worker Queue Configuration into separate config test files.

* CAF-2108 Removed now redundant distribution management section.

* CAF-2108 Adding missing file for testing Job Tracking Worker