SCMOD-12730: Add job pause, resume and status endpoints
https://portal.digitalsafe.net/browse/SCMOD-12730

Adding 3 new endpoints to pause, resume, and get the status of the job. The pause and resume endpoints currently just update the status of the job in the job service database.

Future PRs will cover additional updates to store and recover messages when a job is paused and resumed. See the Jira linked above for full implementation details.